devops/01-python-flask-api/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.27
Werkzeug==2.3.7